fix: address PR review blocking issues
1. Migration version ordering fix:
   - Rename 20260720_002 to 20260721_002 for correct chronological order
   - Migration chain: 20260720_001 → 20260721_001 → 20260721_002

2. Code deduplication:
   - Create shared module tenant_config_cache.py
   - Extract _get_tenant_sensitive_keyword_config to shared module
   - Update session_manager.py and llm_proxy_handler.py to use shared cache
